fix(state): preserve schema and Doctor admission ordering
Check schema versions before reading ownership metadata, commit Doctor config migrations before admission diagnostics, and retain raw agent-copy inspection at the established preflight owner.
