+++ src/flows/doctor-health-contributions-initial.ts
@@ -57,6 +57,12 @@ export function resolveInitialDoctorHealthContributions(params: {
 }): DoctorHealthContribution[] {
   return [
     createDoctorHealthContribution({
+      id: "doctor:write-config-migrations",
+      label: "Write config migrations",
+      required: true,
+      run: runInitialConfigWriteHealth,
+    }),
+    createDoctorHealthContribution({
       id: "doctor:agent-database-admission",
       label: "Agent database admission",
       healthChecks: {
@@ -75,12 +81,6 @@ export function resolveInitialDoctorHealthContributions(params: {
           }));
         },
       },
-    }),
-    createDoctorHealthContribution({
-      id: "doctor:write-config-migrations",
-      label: "Write config migrations",
-      required: true,
-      run: runInitialConfigWriteHealth,
     }),
     createDoctorHealthContribution({
       id: "doctor:node-runtime",
